Update menu-management.html
adding functionality for menu items is now working, the add form enable to add as many sizes and cost of a food item as there are (currently its is manually set, but can be easily retrieved from DB).
main problem was to make functionality of size and cost inputs, to increase them as per requirement, and to remove them if needed, while entering the data. All the data structures needed to be modified accordingly.